Test SQL schema
