+++ tests/Sqlite/AssignmentsStorageTest.php
@@ -6,5 +6,5 @@ namespace Yiisoft\Rbac\Cycle\Tests\Sqlite;
 
 class AssignmentsStorageTest extends \Yiisoft\Rbac\Cycle\Tests\Base\AssignmentsStorageTest
 {
-    use SqliteTrait;
+    use DatabaseTrait;
 }
