+++ tests/Sqlite/ItemsStorageTest.php
@@ -6,5 +6,5 @@ namespace Yiisoft\Rbac\Cycle\Tests\Sqlite;
 
 class ItemsStorageTest extends \Yiisoft\Rbac\Cycle\Tests\Base\ItemsStorageTest
 {
-    use SqliteTrait;
+    use DatabaseTrait;
 }
